+++ src/pretalx/orga/forms/review.py
@@ -1,6 +1,7 @@
 from django import forms
 from django.utils.translation import ngettext_lazy, ugettext_lazy as _
 
+from pretalx.common.forms.widgets import MarkdownWidget
 from pretalx.common.mixins.forms import ReadOnlyFlag
 from pretalx.common.phrases import phrases
 from pretalx.submission.models import Review
@@ -86,3 +87,6 @@ class ReviewForm(ReadOnlyFlag, forms.ModelForm):
     class Meta:
         model = Review
         fields = ('text', 'score')
+        widgets = {
+            'text': MarkdownWidget,
+        }